requirements.txt
langchain>=0.1.14
langchain-community>=0.0.21
langchain-core>=0.1.41
langchain-groq>=0.0.2
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
python-dotenv>=1.0.1
streamlit>=1.35.0
